Sunflower vs Rose/requirements.txt

# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0

# Data Processing
scikit-learn>=1.2.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars
tqdm>=4.65.0

# Optional: For Jupyter notebooks
# jupyter>=1.0.0
# ipywidgets>=8.0.0

# Optional: For model interpretability
# grad-cam>=1.4.0
# captum>=0.6.0
